Consolidate error files
